requirements.txt
Babel==2.6.0
Django==2.1.8
django-appconf==1.0.3
django-compressor==2.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==1.6.7
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
factory-boy==2.11.1
Faker==1.0.5
mock==2.0.0
pbr==5.2.0
phonenumbers==8.10.10
Pillow==6.0.0
psycopg2==2.8.2
purl==1.5
python-dateutil==2.8.0
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.12.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
Unidecode==1.0.23
